// src/app/modules/accounting.luxuryapp/general-ledger/financial-reports/interfaces/aspel-budget.interface.ts
export interface IFlujoCajaDto {
  nombreEmpresa: string;
  periodoPresupuesto: string;
  columnas: string[];
  grupos: IFlujoCajaGrupoDto[];
}

export interface IFlujoCajaGrupoDto {
  nombre: string;
  filas: IFlujoCajaFilaDto[];
  filaTotales?: IFlujoCajaFilaDto;
}

export interface IFlujoCajaFilaDto {
  concepto: string;
  signo: string;
  montos: number[];
  montosMtto: number[];
  montosObrasMejoras: number[];
  montosFondoReserva: number[];
  esSuma: boolean;
  esResta: boolean;
  esFilaTotal: boolean;
  esEncabezado: boolean;
  ocultarOrigen: boolean;
  esManual: boolean;
  origen: string;
}